Refactor(#61): 댓글 관련 폴더 구조 변경
